api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "CSCI 3461 Group 5" }]
dependencies = [
    "flask>=3.0.2",
    "python-dotenv>=1.0.1",
    "mysql-connector-python>=8.3.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "flask --app src.api.app --debug run"
start = "flask --app src.api.app run"
